// frontend/src/app/core/domain/dto/dialog.dto.ts
export interface InformationDialogDTO {
    asset?: string;
    assetClass?: string;
    icon?: string;
    iconClass?: string;
    title: string;
    titleClass?: string;
    subtitle: string;
    subtitleClass?: string;
    submitButton: string;
    submitClass?: string;
    cancelButton?: string;
    cancelClass?: string;
}